Move createOptions() function to a new component called create-options. #3
